+++ braintree/__init__.py
@@ -6,7 +6,7 @@ from error_codes import ErrorCodes
 from error_result import ErrorResult
 from errors import Errors
 from environment import Environment
-from paged_collection import PagedCollection
+from resource_collection import ResourceCollection
 from search import Search
 from subscription import Subscription
 from transaction import Transaction