+++ braintree/credit_card.py
@@ -38,6 +38,10 @@ class CreditCard(Resource):
 
         print(result.credit_card.token)
         print(result.credit_card.masked_number)
+
+    For more information on CreditCards, see http://www.braintreepaymentsolutions.com/gateway/credit-card-api
+
+    For more information on CreditCard verifications, see http://www.braintreepaymentsolutions.com/gateway/credit-card-verification-api
     """
 
     @staticmethod