+++ braintree/transaction.py
@@ -8,7 +8,7 @@ from braintree.address import Address
 from braintree.configuration import Configuration
 from braintree.credit_card import CreditCard
 from braintree.customer import Customer
-from braintree.paged_collection import PagedCollection
+from braintree.resource_collection import ResourceCollection
 from braintree.transparent_redirect import TransparentRedirect
 from braintree.exceptions.not_found_error import NotFoundError
 
@@ -61,6 +61,8 @@ class Transaction(Resource):
 
         print(result.transaction.amount)
         print(result.transaction.order_id)
+
+    For more information on Transactions, see http://www.braintreepaymentsolutions.com/gateway/transaction-api
     """
 
     class Type(object):
@@ -207,7 +209,7 @@ class Transaction(Resource):
         """
         Search for transactions based on keywords. For example, you can search for
         transactions based on the credit card bin. The search will return a
-        :class:`PagedCollection <braintree.paged_collection.PagedCollection>`::
+        :class:`ResourceCollection <braintree.resource_collection.ResourceCollection>`::
 
             collection = braintree.Transaction.search("411111")
 
@@ -222,7 +224,7 @@ class Transaction(Resource):
 
         query_string = urllib.urlencode([("q", query), ("page", page)])
         response = Http().get("/transactions/all/search?" + query_string)
-        return PagedCollection(query, response["credit_card_transactions"], Transaction)
+        return ResourceCollection(query, response["credit_card_transactions"], Transaction)
 
     @staticmethod
     def submit_for_settlement(transaction_id, amount=None):
@@ -322,7 +324,7 @@ class Transaction(Resource):
             "amount", "customer_id", "order_id", "payment_method_token", "type",
             {
                 "credit_card": [
-                    "token", "cvv", "expiration_date", "number"
+                    "token", "cardholder_name", "cvv", "expiration_date", "expiration_month", "expiration_year", "number"
                 ]
             },
             {