+++ braintree/validation_error_collection.py
@@ -1,7 +1,11 @@
 from braintree.validation_error import ValidationError
 
 class ValidationErrorCollection(object):
-    """A class representing a collection of validation errors."""
+    """
+    A class representing a collection of validation errors.
+
+    For more information on ValidationErrors, see http://www.braintreepaymentsolutions.com/gateway/validation-errors
+    """
 
     def __init__(self, data={"errors": []}):
         self.data = data